% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_intro_exp.R
\name{intro_tab_div2}
\alias{intro_tab_div2}
\title{This is the user interface for the intro pages bfore testing}
\usage{
intro_tab_div2()
}
\description{
It looks a little crazy, but is relatively straightforward.
}
