% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnt-package.R
\docType{package}
\name{pnt-package}
\alias{pnt}
\alias{pnt-package}
\title{pnt: Computer adaptive forms of the Philadelphia Naming Test}
\description{
A Shiny web-app for administering and scoring the computer adaptive Philadelphia Naming Test.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rbcavanaugh.github.io/pnt/}
  \item \url{https://github.com/rbcavanaugh/pnt}
  \item Report bugs at \url{https://github.com/rbcavanaugh/pnt/issues}
}

}
\author{
\strong{Maintainer}: Robert Cavanaugh \email{rob.cavanaugh@pitt.edu}

Authors:
\itemize{
  \item Alexander Swiderski
  \item Gerasimos Fergadiotis
  \item William D. Hula
}

}
\keyword{internal}
